pyproject.toml
[tool.poetry]
name = "web_monitoring"
version = "0.1.0"
description = "Small utility for monitoring website changes"
authors = ["JoshKing56 <joshuaiking56@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
click = "^8.0.3"
beautifulsoup4 = "^4.10.0"
urllib3 = "^1.26.7"
telegram-send = "^0.25"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
